+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # Unreleased
 
+# 2025-03-12
+## Added
+- clojure: parinfer group and toggle key map
+- practicalli: which-key sub-menu for Visual-Multi Cursors (Multiple Cursors)
+
 ## Changes
 - lsp: revert to using mason install of Clojure LSP (Mason Registry updated)
 - practicalli: simplify conditional config expressions, remove local whichkey (redundant)
+- clojure: update names for conjure sub-menus
+- readme: revert to using mason for clojure lsp, link to local server config example
 
 # 2024-12-17 Local install of Clojure LSP